requirements.txt
# Database dependencies
alembic==1.13.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Vector database
qdrant-client==1.7.0

# Development dependencies
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0